// src/types.tsx
export interface CategoryDataType {
    id: number;
    name: string;
    image: string;
    parent_category_id: null;
    description: string | null;
    status: boolean;
    slug: string;
    subcategories_count: number;
    parent_category_count: number;
    courses_count: number;
    has_subcategory: boolean;
    has_parent_category: boolean;
    subcategories?: SubCategoriesDataType[];
}

export interface SubCategoriesDataType {
    id: number;
    sequence: null;
    name: string;
    image: string;
    parent_category_id: number;
    description: null;
    status: boolean;
    slug: string;
    created_at: string;
    updated_at: string;
    deleted_at: null;
    has_subcategory: boolean;
    has_parent_category: boolean;
}
// ...
export interface SidebarFilterTypes {
    level: string;
    language: string;
    duration: string | number; // Can be a number (single selection) or string (comma-separated for multiple selections)
    price: string;
    rating: string;
    category: string;
}